# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OverUnder API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/relayer/jobs/{job_id}": {
      "get": {
        "tags": ["relayer"],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_relayer_jobs__job_id__get",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Job Id" } },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/relayer/jobs": {
      "get": {
        "tags": ["relayer"],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_relayer_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Status" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "maximum": 200, "minimum": 1, "default": 50, "title": "Limit" }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/relayer/status": {
      "get": {
        "tags": ["relayer"],
        "summary": "Status",
        "operationId": "status_api_v1_relayer_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/relayer/tick": {
      "post": {
        "tags": ["relayer"],
        "summary": "Tick",
        "description": "One worker pass: the manual / serverless drain when the background loop is off.",
        "operationId": "tick_api_v1_relayer_tick_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/api/v1/ramps/moonpay/session": {
      "post": {
        "tags": ["ramps"],
        "summary": "Moonpay Session",
        "description": "Sign MoonPay widget URL with secret for the authenticated user's address.\n\nEnforces KYC gate: returns 403 if notional ≥ threshold or restricted jurisdiction\nand KYC status not in {pass, not_required}.",
        "operationId": "moonpay_session_api_v1_ramps_moonpay_session_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MoonPaySessionRequest" } } }
        },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/api/v1/ramps/onramp-url": {
      "get": {
        "tags": ["ramps"],
        "summary": "Onramp Url",
        "operationId": "onramp_url_api_v1_ramps_onramp_url_get",
        "parameters": [
          { "name": "address", "in": "query", "required": true, "schema": { "type": "string", "title": "Address" } },
          {
            "name": "usdc_amount",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "100", "title": "Usdc Amount" }
          }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/ramps/offramp-url": {
      "get": {
        "tags": ["ramps"],
        "summary": "Offramp Url",
        "operationId": "offramp_url_api_v1_ramps_offramp_url_get",
        "parameters": [
          { "name": "address", "in": "query", "required": true, "schema": { "type": "string", "title": "Address" } }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoonPaySessionRequest": {
        "properties": { "usdc_amount": { "type": "string", "title": "Usdc Amount", "default": "100" } },
        "type": "object",
        "title": "MoonPaySessionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}